fix: benefit money 타입 double로 변경

+++ spring-test-concurrency-coupon/src/main/java/coupon/domain/Benefit.java
@@ -25,5 +25,5 @@ class Benefit {
     private Member member;
 
     @Column(nullable = false)
-    private Long money;
+    private Double money;
 }
